Fixed node-node and node-edge forces being able to trigger infinite regress of edge adjustment if a node was too large for its neighbour's edges to be unstressed (it's own were already not included in these forces.
Did this by tracking desired average lengths of all node-node shortest paths through the graph and when I need a node-node or a node-edge radius, using the smaller of what the two node radii want (r1 + r2) and what the paths want (unstressed edges).

Added experiment with "void" nodes which are big and invisible, not intended to expand into rooms and serve only to make holes in the layout.