Prototyped a door/key mechanism.
Added a template post-expand mechanism to enable the system fro rename inserted door/key pairs to match one-another (eventually this may add semantics/gameplay mechanisc in the same way).
- Made node names mutatable through INode to allow renaming...

Commented out templates that aren't yet compatible with this approach to doors/keys (anything that makes a loop) and also a few that are compatible but which diluted the attempt to test doors.

Added check in TryAllNodesExpandStepper so that if we are down to the last expandable node, and we haven't reached the target size, we try only further-expandable templates against it...
- This mechanism may need some further elaboration later, as it doesn't take into account that some non-expandable templates might expand the graph enough for completion
- ... or that if we have more than one set of template/non-terminal symbol, then what is available at some place and time may not be expandable anyway...

Replaces "FilterByCodes" with collection.stream() code (as I need it on collections of different types now, and I wantred to try stream() anyway...